+++ resources/js/Components/GlobalSearch/index.ts
@@ -1,0 +1,3 @@
+// resources/js/Components/GlobalSearch/index.ts
+
+export { default as GlobalSearchModal } from "./GlobalSearchModal";